# Speedup algorithm used by submodel_names