I profiled some code looking for ways to speed up modeling and saw
that a good chunk of time was spent in submodel_names. To speed it up,
I rewrote the code to use a new algorithm that only makes one pass
over the list of names.

Supersede #6115 and #6064