@astrofrog 2017-05-25
@bernie-simon - if possible, please do not re-open new pull requests every time. You should be able to force push to the original branch. We prefer to keep everything in the same pull request to keep the conversation in one place.