@pllim 2017-05-25
@bsipocz , I sat down with Bernie, cleaned up his fork, and helped him create this PR, so it should be okay now.