@bsipocz 2017-05-25
I would say that if this is a significant speed-up, a mention of it would be nice in the changelog.