@pllim 2017-05-25
@bernie-simon , did you profile before and after?