@nden 2017-05-25
There were changes since this work was started. It's best to profile now.