@bernie-simon 2017-05-25
Ideally I would have a file with a complex gwcs model to test this. I have several, but they no longer seem to be valid, datamodels crashes when I try to open them. If you have such a file, could you let me know where it is?