@bernie-simon 2017-05-25
OK, here are the first results. The script opens and then closes all the reference files I have from crds - twenty times. Timed with the Unix time utility:

Without the change:

* real	2m53.616s
* user	2m4.340s
* sys	0m41.177s

With the change:

* real	2m30.904s
* user	1m56.333s
*  sys	0m34.016s