@pllim 2017-05-25
From a user's standpoint, does 2m54s cutting down to 2m31s warrant a change log, @bsipocz and @nden ?