@nden 2017-05-25
Oh, never mind. You are not timing the `Model.submodel_names` but opening and closing reference files?