@nden 2017-05-25
Here's my timing of only the change using `timeit Model.submodel_names`  on 109 models.

before:
```
timeit w1.forward_transform.submodel_names
587 ms ± 1.73 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
after
```
timeit w1.forward_transform.submodel_names
588 ms ± 1.59 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

```