@bernie-simon 2017-05-26
I'm still testing. Testing was held up yesterday by having an old version of asdf on my machine. The results are naturally going to depend on how much the affected method is called.