@bernie-simon 2017-05-26
OK, here is the test case Nadia sent me, run in a loop 20 times:

Original code

* real	1m54.704s
* user	2m5.537s
* sys	0m1.239s

With my change

* real	1m35.054s
* user	1m45.862s
* sys	0m1.243s